LayoutTests/media/modern-media-controls/buttons-container/buttons-container-buttons-property.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/buttons-container.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/buttons-container.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing <code>ButtonsContainer</code> buttons property.");

window.jsTestIsAsync = true;

const container = new ButtonsContainer({
    margin: 10,
    padding: 20
});

container.buttons = [new Button, new Button, new Button];

scheduler.frameDidFire = function()
{
    shouldBe("container.buttons.length", "3");
    shouldBe("container.children.length", "3");
    shouldBe("container.buttons", "container.children");
    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
