LayoutTests/media/modern-media-controls/buttons-container/buttons-container-buttons-property-expected.txt

Testing ButtonsContainer buttons property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS container.buttons.length is 3
PASS container.children.length is 3
PASS container.buttons is container.children
PASS successfullyParsed is true

TEST COMPLETE
